optimizations to shader inputs/data processing and anti-ball stretch code
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2412 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ vpinball.cpp
@@ -8823,8 +8823,8 @@ INT_PTR CALLBACK DimensionProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lP
          if (idx >= DIM_TABLE_SIZE || idx < 0)
             break;
 
-         int width = (int)floor(dimTable[idx].width*47.0f + 0.5f);
-         int height = (int)floor(dimTable[idx].height*47.0f + 0.5f);
+         int width = (int)floorf(dimTable[idx].width*47.0f + 0.5f);
+         int height = (int)floorf(dimTable[idx].height*47.0f + 0.5f);
          char textBuf[32];
          sprintf_s(textBuf, "%i", width);
          SetDlgItemText(hwndDlg, IDC_VP_WIDTH, textBuf);
@@ -8854,7 +8854,7 @@ INT_PTR CALLBACK DimensionProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lP
             ret = sscanf_s(textBuf, "%f", &sizeWidth);
             if (ret != 1 || sizeWidth < 0.0f)
                sizeWidth = 0;
-            int width = (int)floor(sizeWidth*47.0f + 0.5f);
+            int width = (int)floorf(sizeWidth*47.0f + 0.5f);
             sprintf_s(textBuf, "%i", width);
             SetDlgItemText(hwndDlg, IDC_VP_WIDTH, textBuf);
          }
@@ -8865,7 +8865,7 @@ INT_PTR CALLBACK DimensionProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lP
             ret = sscanf_s(textBuf, "%f", &sizeHeight);
             if (ret != 1 || sizeHeight < 0.0f)
                sizeHeight = 0;
-            int height = (int)floor(sizeHeight*47.0f + 0.5f);
+            int height = (int)floorf(sizeHeight*47.0f + 0.5f);
             sprintf_s(textBuf, "%i", height);
             SetDlgItemText(hwndDlg, IDC_VP_HEIGHT, textBuf);
          }
